requirements.txt
alpaca-trade-api
python-dotenv
requests
pandas
numpy
matplotlib
